requirements.txt
Django==2.0.2
pytz==2018.3
djangorestframework==3.7.7